requirements.txt
ultralytics==8.0.196
streamlit==1.28.0
pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
pandas==2.0.3
torch==2.0.1
torchvision==0.15.2
protobuf==3.20.3
```

### 2. **`runtime.txt`**
```
python-3.11.7